+++ code/ahandle.h
@@ -15,6 +15,8 @@
 
 #include "vqaplay.h"
 
+#include <cstdint>
+
 #ifdef _WIN32
 #include <dsound.h>
 #else
@@ -78,10 +80,10 @@ struct Ahandle {
 unsigned int __cdecl Simple_Timer_Callback_Audio_Handler(VQAHandle *vqa);
 unsigned int __cdecl Timer_Callback_Audio_Handler(VQAHandle *vqa);
 
-long __cdecl Lock_Audio_Handler(void);
-long __cdecl Unlock_Audio_Handler(void);
-intptr_t __cdecl Stream_Audio_Handler(VQAHandle *vqa, long action, void *buffer, long nbytes);
+int32_t __cdecl Lock_Audio_Handler(void);
+int32_t __cdecl Unlock_Audio_Handler(void);
+intptr_t __cdecl Stream_Audio_Handler(VQAHandle *vqa, uint32_t action, void *buffer, int32_t nbytes);
 
 
-typedef long (__cdecl *AHANDLE_CALLBACK_1)(VQAHandle *vqa);
-typedef long (__cdecl *AHANDLE_CALLBACK_2)(VQAHandle *vqa, void *buffer);
+typedef int32_t (__cdecl *AHANDLE_CALLBACK_1)(VQAHandle *vqa);
+typedef int32_t (__cdecl *AHANDLE_CALLBACK_2)(VQAHandle *vqa, void *buffer);
