+++ code/vqalib/dstream.cpp
@@ -54,7 +54,7 @@
 #include	<string.h>
 
 
-intptr_t __cdecl Disk_VQA_Stream_Handler(VQAHandle *vqa, long action, void *buffer, long nbytes)
+intptr_t __cdecl Disk_VQA_Stream_Handler(VQAHandle *vqa, uint32_t action, void *buffer, int32_t nbytes)
 {
 	long fh;
 	long error = 0;
@@ -157,7 +157,7 @@ intptr_t __cdecl Disk_VQA_Stream_Handler(VQAHandle *vqa, long action, void *buff
 
 
 
-intptr_t __cdecl Memory_VQA_Stream_Handler(VQAHandle *vqa, long action, void *buffer, long nbytes)
+intptr_t __cdecl Memory_VQA_Stream_Handler(VQAHandle *vqa, uint32_t action, void *buffer, int32_t nbytes)
 {
 	long error = 0;
 	int p;
